internal/c2go: fix integer overflow by converting coordinate types to int32
This commit resolves a critical infinite loop issue in _adjustPentVertOverage
by converting coordinate-related types from Go's platform-dependent int (64-bit)
to int32 to match H3 C library's 32-bit int behavior exactly.

Root cause:
- H3 C uses 32-bit int, causing coordinate sum 2147483652 to wrap to -2147483644
- Go used 64-bit int, keeping 2147483652 positive, causing infinite overage loop
- Critical case: Face=5, Coord={715827882,0,1431655770} at res=3

Changes:
- Convert CoordIJK.{I,J,K} and CoordIJ.{I,J} to int32
- Update BaseCellData.CwOffsetPent to [2]int32
- Update BaseCellRotation.{BaseCell,CcwRot60} to int32
- Update CellBoundary.NumVerts to int32
- Add int32/int casting at function boundaries
- Keep boolean semantics as bool for clarity

Testing:
- Critical difference test now passes (both Go/C return overage=0)
- Pentagon vertex overage completes in ~24s (no infinite loop)
- Full c2go test suite passes

+++ internal/c2go/baseCells___baseCellIsCwOffset.go
@@ -4,6 +4,6 @@ package c2go
 // For non-pentagon base cells, this always returns false.
 // Ported from H3 C: baseCells.c::_baseCellIsCwOffset
 func _baseCellIsCwOffset(baseCell int, testFace int) bool {
-	return baseCellData[baseCell].CwOffsetPent[0] == testFace ||
-		baseCellData[baseCell].CwOffsetPent[1] == testFace
+	return baseCellData[baseCell].CwOffsetPent[0] == int32(testFace) ||
+		baseCellData[baseCell].CwOffsetPent[1] == int32(testFace)
 }
